requirements.txt
# Core Dependencies
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.20
requests>=2.31.0

# Document Processing
pypdf>=3.17.0

# Vector Store & Embeddings
chromadb>=0.4.15
sentence-transformers>=2.2.2

# Additional utilities
python-dotenv>=1.0.0
typing-extensions>=4.8.0

# Optional: For better performance
numpy>=1.24.0
pandas>=2.1.0
